tutorial/24_data_extraction_pipeline_with_adk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk-data-extraction"
version = "0.1.0"
description = "Google ADK structured data extraction examples and tutorials"
authors = [{name = "ADK Tutorial", email = "tutorial@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["google", "adk", "ai", "data-extraction", "llm", "structured-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "PyMuPDF>=1.23.0", # For PDF processing
    "python-multipart>=0.0.6", # For file uploads
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "google-adk>=1.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "pyright>=1.1.350",
    "pre-commit>=3.6.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

jupyter = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.29.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.17.0",
]

[project.scripts]
adk-extract-contacts = "adk_data_extraction.examples.basic_contact_extraction:main"
adk-analyze-legal = "adk_data_extraction.examples.legal_document_analysis:main"

[project.urls]
Homepage = "https://github.com/your-org/adk-data-extraction"
Documentation = "https://your-org.github.io/adk-data-extraction"
Repository = "https://github.com/your-org/adk-data-extraction"
"Bug Tracker" = "https://github.com/your-org/adk-data-extraction/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_data_extraction"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.ruff.isort]
known-first-party = ["adk_data_extraction"]

[tool.pytest.ini_options]
testpaths = ["tests"] 
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=adk_data_extraction",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["adk_data_extraction"]
omit = [
    "*/tests/*",
    "*/examples/*",
    "*/notebooks/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "All"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
